Reorganização dos códigos. Todos os módulos importáveis foram movidos para o pacote `lib`, e apenas os executáveis estão no top-level.
