+++ balaio/lib/validations/__init__.py
@@ -1,0 +1,11 @@
+# This relies on each of the submodules having an __all__ variable.
+from .base import *
+from .article import *
+from .journal import *
+from .references import *
+
+__all__ = (base.__all__ +
+           article.__all__ +
+           journal.__all__ +
+           references.__all__)
+
