+++ balaio/tests/utils.py
@@ -1,7 +1,7 @@
 from sqlalchemy import create_engine
 from sqlalchemy.exc import OperationalError
 
-from balaio import models
+from balaio.lib import models
 
 
 DB_NAME = 'app_balaio_tests'
@@ -18,7 +18,7 @@ def db_bootstrap():
 
     :returns: an instance of engine.
     """
-    engine = create_engine('postgresql+psycopg2://postgres:@localhost/%s' % DB_NAME, echo=False)
+    engine = create_engine('postgresql+psycopg2://postgres:postgres@localhost/%s' % DB_NAME, echo=False)
     models.Base.metadata.drop_all(engine)
     models.init_database(engine)
